requirements.txt
torch
torchviz
# simpletransformers
openai-whisper
# transformers
numba
numpy
pandas
matplotlib
scipy
torchaudio
tqdm
scikit-learn
wavio
kapre
munch
SoundFile
pydub
pyyaml
librosa
phonemizer
sounddevice
SpeechRecognition
# git+https://github.com/resemble-ai/monotonic_align.git
wikipedia-api~=0.5.4
pyttsx3~=2.90
vosk-tts~=0.3.45
termcolor~=1.1.0
pyautogui
requests==2.30.0
audioplayer
python-dateutil~=2.8.2
playsound
addict
sonopy
sox
pygame
ffmpeg
deep_utils
pytorch_lightning
torchmetrics